requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic>=0.40.0
openai>=1.0.0
google-genai==1.14.0
supabase==2.28.3
python-dotenv==1.0.1
httpx==0.28.1
pydantic>=2.10.4
